benchdnn: make desc_t.ndims of int type

+++ tests/benchdnn/bnorm/bnorm.hpp
@@ -48,7 +48,7 @@ struct desc_t {
     int64_t mb, ic, id, ih, iw;
     float eps;
     const char *name;
-    int64_t ndims;
+    int ndims;
 };
 int str2desc(desc_t *desc, const char *str);
 std::ostream &operator<<(std::ostream &s, const desc_t &d);
